AnimatedVisualPlayer without App or CoreWindow. (#9047)
## Description
AnimatedVisualPlayer's constructor currently crashes if there isn't an App or CoreWindow available. This change allows it to be created in scenarios where those don't exist.

Note: The events which would normally be registered on App and CoreWindow won't be registered in this case. UI experiences may need to account for that.

## How Has This Been Tested?
Tested in a scenario without App or CoreWindow which no longer crashes and renders correctly with this fix.